COMMERCE-5659 - Consistency
