+++ modules/apps/commerce/commerce-product-service/src/main/java/com/liferay/commerce/product/internal/data/source/AssetCategoriesCPDataSourceImpl.java
@@ -15,7 +15,6 @@
 package com.liferay.commerce.product.internal.data.source;
 
 import com.liferay.asset.kernel.model.AssetEntry;
-import com.liferay.asset.kernel.model.AssetTag;
 import com.liferay.asset.kernel.service.AssetEntryLocalService;
 import com.liferay.commerce.product.catalog.CPQuery;
 import com.liferay.commerce.product.data.source.CPDataSource;
@@ -25,7 +24,6 @@ import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.util.Portal;
 
-import java.util.List;
 import java.util.Locale;
 
 import org.osgi.service.component.annotations.Component;
@@ -36,19 +34,18 @@ import org.osgi.service.component.annotations.Reference;
  */
 @Component(
 	enabled = false, immediate = true,
-	property = "commerce.product.data.source.name=" + CPDataSourceAssetCategoriesAndTagsImpl.NAME,
+	property = "commerce.product.data.source.name=" + AssetCategoriesCPDataSourceImpl.NAME,
 	service = CPDataSource.class
 )
-public class CPDataSourceAssetCategoriesAndTagsImpl
-	extends BaseCPDataSourceAssetEntryImpl {
+public class AssetCategoriesCPDataSourceImpl
+	extends BaseAssetEntryCPDataSourceImpl {
 
-	public static final String NAME = "assetCategoriesAndTagsDataSource";
+	public static final String NAME = "assetCategoriesDataSource";
 
 	@Override
 	public String getLabel(Locale locale) {
 		return LanguageUtil.get(
-			getResourceBundle(locale),
-			"products-of-the-same-categories-and-tags");
+			getResourceBundle(locale), "products-of-the-same-categories");
 	}
 
 	@Override
@@ -64,23 +61,8 @@ public class CPDataSourceAssetCategoriesAndTagsImpl
 			CPDefinition.class.getName(), cpDefinitionId);
 
 		cpQuery.setAnyCategoryIds(assetEntry.getCategoryIds());
-		cpQuery.setAnyTagIds(_getTagIds(assetEntry));
 
 		return cpQuery;
-	}
-
-	private long[] _getTagIds(AssetEntry assetEntry) throws PortalException {
-		List<AssetTag> assetTags = assetEntry.getTags();
-
-		long[] tagIds = new long[assetTags.size()];
-
-		for (int i = 0; i < assetTags.size(); i++) {
-			AssetTag assetTag = assetTags.get(i);
-
-			tagIds[i] = assetTag.getTagId();
-		}
-
-		return tagIds;
 	}
 
 	@Reference(unbind = "-")
