+++ modules/apps/commerce/commerce-product-service/src/main/java/com/liferay/commerce/product/internal/data/source/BaseAssetEntryCPDataSourceImpl.java
@@ -39,7 +39,7 @@ import javax.servlet.http.HttpServletRequest;
 /**
  * @author Ethan Bustad
  */
-public abstract class BaseCPDataSourceAssetEntryImpl implements CPDataSource {
+public abstract class BaseAssetEntryCPDataSourceImpl implements CPDataSource {
 
 	@Override
 	public CPDataSourceResult getResult(
